+++ R/script.R
@@ -35,10 +35,13 @@
 #'   Docker image or an image available on DockerHub. Default image is
 #'   \code{r-base:{jetty:::r_version()}} where your R version is determined from
 #'   your local R session.
-#' @param stdout,stderr Where output to ‘stdout’ or ‘stderr’ should be sent.
-#'   Possible values are "" (send to the R console; the default), NULL or FALSE
-#'   (discard output), TRUE (capture the output in a character vector) or a
-#'   character string naming a file. See \code{\link{system2}} for more details.
+#' @param stdout,stderr where output to ‘stdout’ or ‘stderr’ should be sent.
+#'   Possible values are "", to the R console (the default), NULL
+#'   (discard output), FALSE (discard output), TRUE
+#'   (capture the output silently and then discard), or a
+#'   character string naming a file. See \code{\link{system2}} which this
+#'   function uses under the hood; however, note that \code{\link{system2}}
+#'   handles these options slightly differently.
 #' @param install_dependencies A boolean indicating whether jetty should
 #'   discover packages used in your code and try to install them in the
 #'   Docker container prior to executing the provided function/expression.
@@ -49,7 +52,11 @@
 #'   By default jetty will look for files called ".Rprofile" and ".Renviron"
 #'   in the current working directory. If either file is found, they will be
 #'   transferred to the Docker sub-process and loaded before executing any
-#'   R commands.
+#'   R commands. To explicitly exclude either file, set the value to `NULL`.
+#'   Alternatively, to exclude either file for all jetty function calls,
+#'   set the `JETTY_IGNORE_RPROFILE`/`JETTY_IGNORE_RENVIRON` environment
+#'   variable(s) to one of `c(TRUE, "T")` or set the R option(s)
+#'   `jetty.ignore.rprofile`/`jetty.ignore.renviron` to `TRUE`.
 #' @param debug A boolean indicating whether to print out the commands that are
 #'   being executed via the shell. This is mostly helpful to see what is
 #'   happening when things start to error.
@@ -73,6 +80,12 @@
 #'   context = here::here(),
 #'   install_dependencies = TRUE
 #' )
+#' 
+#' # Execute a script and explicitly ignore an existing .Rprofile
+#' run_script(
+#'   file = here::here("code/analysis_script.R"),
+#'   r_profile = NULL
+#' )
 #' }
 #'
 #' @export
@@ -84,8 +97,8 @@ run_script <- function(
   stdout = "",
   stderr = "",
   install_dependencies = FALSE,
-  r_profile = file.path(getwd(), ".Rprofile"),
-  r_environ = file.path(getwd(), ".Renviron"),
+  r_profile = jetty_r_profile(),
+  r_environ = jetty_r_environ(),
   debug = FALSE
 ) {
   file <- normalizePath(file, mustWork = TRUE)
@@ -102,7 +115,7 @@ run_script <- function(
   on.exit(if (file.exists(temp_out_local)) file.remove(temp_out_local), add = TRUE)
   saveRDS(object = args, file = temp_in_local)
   # Capture the expression to be evaluated
-  expr <- rlang::expr({ setwd(!!context); source })
+  expr <- rlang::expr({ source })
   # If requested, write the code to an R file and examine for dependencies
   dependencies <- take_stock(
     expr = file,
@@ -125,7 +138,8 @@ run_script <- function(
     renv = renv_load,
     temp_out = temp_out_docker,
     temp_in = temp_in_docker,
-    dependencies = dependencies
+    dependencies = dependencies,
+    context = context
   )
   # Generate additional arguments to pass to `docker run ...`
   temp_dir_mount <- bindmount_temp(temp_dir, jetty_temp_dir())