+++ tests/testthat/test-docker.R
@@ -12,9 +12,9 @@ test_that("jetty executes commands and gets expected results", {
   skip_on_cran()
 
   # lm test
-  expect_s3_class(jetty::run({ lm(mpg ~ ., data = mtcars) }, image = "r-base:latest"), class = "lm")
-  expect_s3_class(jetty::run(function(d) lm(mpg ~ ., data = d), args = list(d = mtcars), image = "r-base:latest"), class = "lm")
-  expect_error(jetty::run(function(d) lm(mpg ~ ., data = d), args = c(mtcars), image = "r-base:latest"))
+  expect_s3_class(jetty::run({ lm(mpg ~ ., data = mtcars) }, image = "r-base:latest", r_profile = NULL), class = "lm")
+  expect_s3_class(jetty::run(function(d) lm(mpg ~ ., data = d), args = list(d = mtcars), image = "r-base:latest", r_profile = NULL), class = "lm")
+  expect_error(jetty::run(function(d) lm(mpg ~ ., data = d), args = c(mtcars), image = "r-base:latest", r_profile = NULL))
 
   # using a package (Matrix)
   expect_equal(
@@ -22,7 +22,7 @@ test_that("jetty executes commands and gets expected results", {
       library(Matrix)
       set.seed(123)
       rsparsematrix(10, 10, density = 1)
-    }, image = "r-base:latest"),
+    }, image = "r-base:latest", r_profile = NULL),
     expected = {
       set.seed(123)
       library(Matrix)
@@ -36,13 +36,14 @@ test_that("jetty executes commands and gets expected results", {
       library(Matrix)
       set.seed(123)
       rsparsematrix(10, 10, density = 1)
-    }, image = "r-base:latest"),
+    }, image = "r-base:latest", r_profile = NULL),
     expected = jetty::run({
       set.seed(123)
       function(ncol) Matrix::rsparsematrix(10, ncol, 1)
     },
     args = list(ncol = 10), 
-    image = "r-base:latest")
+    image = "r-base:latest",
+    r_profile = NULL)
   )
 
   # ggplot2 example
@@ -54,7 +55,8 @@ test_that("jetty executes commands and gets expected results", {
         ggplot2::theme_minimal()
     },
     install_dependencies = TRUE,
-    image = "r-base:latest"
+    image = "r-base:latest",
+    r_profile = NULL
   )
   expect_s3_class(plt, class = c("gg", "ggplot"))
 })
@@ -87,4 +89,26 @@ test_that("jetty correctly loads existing .Rprofile and .Renviron", {
       image = "r-base:latest"
     )
   )
+})
+
+test_that("jetty honors R options and System environment variables", {
+
+  if (!interactive()) skip()
+
+  # Set a system environment variable to ignore the .Rprofile
+  cur_val <- Sys.getenv("JETTY_IGNORE_RPROFILE")
+  Sys.setenv("JETTY_IGNORE_RPROFILE" = TRUE)
+  expect_identical(jetty::run(function() var(cars)), var(cars), tolerance = 1e-5)
+  Sys.setenv("JETTY_IGNORE_RPROFILE" = cur_val)
+
+  # Set an R option to ignore the .Rprofile
+  cur_val <- getOption("jetty.ignore.rprofile")
+  options("jetty.ignore.rprofile" = TRUE)
+  expect_identical(jetty::run(function() var(cars)), var(cars), tolerance = 1e-5)
+  options("jetty.ignore.rprofile" = cur_val)
+
+  skip()
+  # Expect an error in this environment because it should fail to
+  # execute the existing .Rprofile
+  expect_error(jetty::run(function() var(cars), stdout = FALSE))
 })